# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Filament DB API",
    "description": "REST API for managing 3D printing filament profiles, nozzles, and printers. Supports filament inheritance (parent/variant), per-printer per-nozzle calibrations, INI profile export/import, and OpenPrintTag NFC binary generation.",
    "version": "1.8.4",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/printers": {
      "get": {
        "tags": ["Printers"],
        "summary": "List printers",
        "description": "Returns all non-deleted printers with populated installed nozzles.",
        "parameters": [
          {
            "name": "manufacturer",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by manufacturer name"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of printers sorted by manufacturer then name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Printer" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Printers"],
        "summary": "Create printer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PrinterInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created printer",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Printer" }
              }
            }
          }
        }
      }
    },
...
    "/api/snapshot": {
      "get": {
        "tags": ["Snapshot"],
        "summary": "Export database snapshot",
        "description": "Downloads a JSON snapshot of the entire database including all filaments, nozzles, and printers.",
        "responses": {
          "200": {
            "description": "JSON snapshot file",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      },
      "post": {
        "tags": ["Snapshot"],
        "summary": "Restore database snapshot",
        "description": "Restores the database from a snapshot using best-effort rollback. If any part of the restore fails, the handler attempts to re-insert previous data from an in-memory backup. Concurrent restores are rejected with 409. Note: the restore is not truly atomic — concurrent readers may observe partial state during the delete/insert window, and rollback itself may fail.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "type": "object", "required": ["file"], "properties": { "file": { "type": "string", "format": "binary" } } }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restore successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "restored": {
                      "type": "object",
                      "properties": {
                        "filaments": { "type": "integer" },
                        "nozzles": { "type": "integer" },
                        "printers": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Restore already in progress", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Restore failed (rolled back)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/tds": {
      "get": {
        "tags": ["TDS"],
        "summary": "Check AI API key status",
        "description": "Returns whether an AI API key is configured and which provider is active.",
        "responses": {
          "200": {
            "description": "Key status",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "configured": { "type": "boolean" }, "provider": { "type": "string", "enum": ["gemini", "claude", "openai"] } } } } }
          }
        }
      },
      "put": {
        "tags": ["TDS"],
        "summary": "Save AI API key",
        "description": "Save and validate an AI API key for TDS extraction.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["apiKey"], "properties": { "apiKey": { "type": "string" }, "provider": { "type": "string", "enum": ["gemini", "claude", "openai"] } } } } }
        },
        "responses": {
          "200": { "description": "Key saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "401": { "description": "Invalid API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["TDS"],
        "summary": "Remove AI API key",
        "description": "Removes the stored API key and resets provider to default.",
        "responses": {
          "200": { "description": "Key removed", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } }
        }
      },
      "post": {
        "tags": ["TDS"],
        "summary": "Extract filament data from TDS",
        "description": "Extract filament properties from a Technical Data Sheet URL or uploaded file using AI.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string" }, "apiKey": { "type": "string" }, "provider": { "type": "string" } } }
            },
            "multipart/form-data": {
              "schema": { "type": "object", "required": ["file"], "properties": { "file": { "type": "string", "format": "binary" }, "apiKey": { "type": "string" }, "provider": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Extraction result", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "description": "API key not configured", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "502": { "description": "Extraction failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "Nozzle": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string", "example": "0.4mm Brass" },
          "diameter": { "type": "number", "example": 0.4 },
          "type": { "type": "string", "example": "Brass" },
          "highFlow": { "type": "boolean", "default": false },
          "hardened": { "type": "boolean", "default": false },
          "notes": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "diameter", "type"]
      },
...
      "Printer": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string", "example": "Prusa Core One" },
          "manufacturer": { "type": "string", "example": "Prusa" },
          "printerModel": { "type": "string", "example": "Core One" },
          "installedNozzles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Nozzle" },
            "description": "Populated nozzle objects"
          },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "manufacturer", "printerModel"]
      },
      "PrinterInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "manufacturer": { "type": "string" },
          "printerModel": { "type": "string" },
          "installedNozzles": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of Nozzle ObjectIds"
          },
          "notes": { "type": "string" }
        },
        "required": ["name", "manufacturer", "printerModel"]
      },
...
    }
  }
}